Commit API Vendedor
Se implementa puerto 8082
se implementa Método buscar vendedor por Id
se implementa Método Eliminar vendedor por Id

+++ vendedores/src/main/java/com/perfumelandiaspa/vendedores/Repository/VendedorRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.perfumelandiaspa.vendedores.Model.Entity.VendedorEntity;
 
-public interface VendedorRepository extends JpaRepository<VendedorEntity, Long>{
+public interface VendedorRepository extends JpaRepository<VendedorEntity, Integer>{
     List<VendedorEntity> findBySucursal(String sucursal); //para el endpoint de sucursal en postman
 
     boolean existsById(int idVendedor);
